refactor: dedupe comparison helpers and harden source titles
